requirements.txt
pyodbc
requests
pillow
